+++ CHANGELOG.md
@@ -17,3 +17,5 @@
 - Add async version of `Prelude.Try`.
 
 - Add async versions of `Result<T>.Match` and `Result<T>.Switch`.
+
+- Rename `Result<T>.ToType` to `To` and allow attempting to convert to non-derived types.